requirements.txt
Django==1.9.2
MySQL-python==1.2.5
Pillow==3.1.0
South==1.0.2
argparse==1.1
boto==2.39.0
certifi==14.05.14
coverage==3.7.1
django-ace==1.0.2
django-annoying==0.8.1
django-autofixture==0.12.0
django-bootstrap3==5.1.1
django-scheduler==0.7.4
django-selectable==0.9.0
django-ses==0.7.0
future==0.14.3
gdata==2.0.18
icalendar==3.9.1
mock==1.0.1
psycopg2==2.6.1
python-dateutil==2.4.1
python-google-spreadsheet==2.0.0
pytz==2014.10
requests==2.9.1
six==1.9.0
wsgiref==0.1.2
